Drop kinderiv dimplic stuff
