+++ coreneuron/sim/scopmath/newton_thread.cpp
@@ -1,58 +1,6 @@
-/******************************************************************************
- *
- * File: newton.c
- *
- * Copyright (c) 1987, 1988, 1989, 1990
- *   Duke University
- *
- ******************************************************************************/
-
-/*------------------------------------------------------------*/
-/*                                                            */
-/*  NEWTON                                        	      */
-/*                                                            */
-/*    Iteratively solves simultaneous nonlinear equations by  */
-/*    Newton's method, using a Jacobian matrix computed by    */
-/*    finite differences.				      */
-/*                                                            */
-/*  Returns: 0 if no error; 2 if matrix is singular or ill-   */
-/*			     conditioned; 1 if maximum	      */
-/*			     iterations exceeded	      */
-/*                                                            */
-/*  Calling sequence: newton(n, x, p, pfunc, value)	      */
-/*                                                            */
-/*  Arguments:                                                */
-/*                                                            */
-/*    Input: n, integer, number of variables to solve for.    */
-/*                                                            */
-/*           x, pointer to array  of the solution             */
-/*		vector elements	possibly indexed by index     */
-/*                                                            */
-/*	     p,  array of parameter values		      */
-/*                                                            */
-/*           pfunc, pointer to function which computes the    */
-/*               deviation from zero of each equation in the  */
-/*               model.                                       */
-/*                                                            */
-/*	     value, pointer to array to array  of             */
-/*		 the function values.			      */
-/*                                                            */
-/*    Output: x contains the solution value or the most       */
-/*               recent iteration's result in the event of    */
-/*               an error.                                    */
-/*                                                            */
-/*  Functions called: makevector, freevector, makematrix,     */
-/*		      freematrix			      */
-/*		      buildjacobian, crout, solve	      */
-/*                                                            */
-/*------------------------------------------------------------*/
-
 #include <math.h>
 #include <stdlib.h>
 
-#include "coreneuron/mechanism/mech/mod2c_core_thread.hpp"
-#include "coreneuron/sim/scopmath/errcodes.h"
-#include "coreneuron/sim/scopmath/newton_struct.h"
 #include "coreneuron/sim/scopmath/newton_thread.hpp"
 #include "coreneuron/utils/nrnoc_aux.hpp"
